+++ Online/Events/ResourceRequest.cs
@@ -3,14 +3,13 @@
     public class ResourceRequest : ResourceEvent
     {
         public ResourceRequest() { }
-
         public ResourceRequest(OnlineResource onlineResource) : base(onlineResource) { }
-
-        public override EventTypeId eventType => EventTypeId.ResourceRequest;
 
         public override void Process()
         {
             onlineResource.Requested(this);
         }
+
+        public override EventTypeId eventType => EventTypeId.ResourceRequest;
     }
 }
