+++ Online/Events/ResourceTransfer.cs
@@ -1,25 +1,9 @@
-﻿using System.Collections.Generic;
-
-namespace RainMeadow
+﻿namespace RainMeadow
 {
     public class ResourceTransfer : ResourceEvent
     {
-        public List<OnlinePlayer> participants;
-        public List<OnlineEntity.EntityId> abandonedEntities;
-
         public ResourceTransfer() { }
-        public ResourceTransfer(OnlineResource resource, List<OnlinePlayer> subscribers, List<OnlineEntity.EntityId> abandonedEntities) : base(resource)
-        {
-            this.participants = subscribers;
-            this.abandonedEntities = abandonedEntities;
-        }
-
-        public override void CustomSerialize(Serializer serializer)
-        {
-            base.CustomSerialize(serializer);
-            serializer.Serialize(ref participants);
-            serializer.Serialize(ref abandonedEntities);
-        }
+        public ResourceTransfer(OnlineResource resource) : base(resource) { }
 
         public override void Process()
         {
