+++ Online/Resource/OnlineResource.Entities.cs
