+++ Online/Resource/WorldSession.Entities.cs
