+++ Rain Meadow.csproj
@@ -161,6 +161,7 @@
     <Compile Include="Menu\IHaveADescription.cs" />
     <Compile Include="Menu\MenuDarkSprite.cs" />
     <Compile Include="Menu\SmartMenu.cs" />
+    <Compile Include="Online\State\EntityState.cs" />
     <Compile Include="Online\OnlineController.cs" />
     <Compile Include="Online\EntityFeed.cs" />
     <Compile Include="Online\Events\EntityReleaseEvent.cs" />
@@ -178,39 +179,42 @@
     <Compile Include="Online\OnlineEntity.cs" />
     <Compile Include="Online\OnlineEntity.EntityId.cs" />
     <Compile Include="Online\OnlineEntity.EntityState.cs" />
-    <Compile Include="Online\OnlineResource.Entities.cs" />
-    <Compile Include="Online\OnlineResource.LeaseState.cs" />
-    <Compile Include="Online\OnlineResource.State.cs" />
-    <Compile Include="Online\OnlineResource.Transactions.cs" />
-    <Compile Include="Online\OnlineState.cs" />
+    <Compile Include="Online\Resource\OnlineResource.Entities.cs" />
+    <Compile Include="Online\Resource\OnlineResource.LeaseState.cs" />
+    <Compile Include="Online\Resource\OnlineResource.State.cs" />
+    <Compile Include="Online\Resource\OnlineResource.Transactions.cs" />
+    <Compile Include="Online\State\OnlineState.cs" />
     <Compile Include="Online\PlayersManager.cs" />
-    <Compile Include="Online\RoomSession.Entities.cs" />
+    <Compile Include="Online\Resource\RoomSession.Entities.cs" />
     <Compile Include="Online\Serializer.cs" />
-    <Compile Include="Online\Lobby.cs" />
+    <Compile Include="Online\Resource\Lobby.cs" />
     <Compile Include="Menu\LobbyInfo.cs" />
     <Compile Include="Online\LobbyManager.cs" />
     <Compile Include="Menu\LobbyMenu.cs" />
     <Compile Include="Online\OnlineManager.cs" />
     <Compile Include="Menu\LobbySelectMenu.cs" />
     <Compile Include="Online\OnlinePlayer.cs" />
-    <Compile Include="Online\OnlineResource.cs" />
+    <Compile Include="Online\Resource\OnlineResource.cs" />
     <Compile Include="Game\OnlineGameSession.cs" />
     <Compile Include="Game\OnlineGameSession.Rules.cs" />
     <Compile Include="Online\Events\OnlineEvent.cs" />
     <Compile Include="Online\Events\ResourceRelease.cs" />
     <Compile Include="Online\Events\ReleaseResult.cs" />
     <Compile Include="Online\Events\ResourceRequest.cs" />
-    <Compile Include="Online\RoomSession.cs" />
+    <Compile Include="Online\Resource\RoomSession.cs" />
     <Compile Include="Online\Events\ResourceTransfer.cs" />
     <Compile Include="Online\Events\TransferResult.cs" />
+    <Compile Include="Online\State\PhysicalObjectEntityState.cs" />
+    <Compile Include="Online\State\RealizedObjectState.cs" />
+    <Compile Include="Online\State\RealizedPlayerState.cs" />
     <Compile Include="Online\Subscription.cs" />
-    <Compile Include="Online\WorldSession.cs" />
-    <Compile Include="Online\WorldSession.Entities.cs" />
+    <Compile Include="Online\Resource\WorldSession.cs" />
+    <Compile Include="Online\Resource\WorldSession.Entities.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RainMeadow.cs" />
     <Compile Include="Game\RainMeadow.GameHooks.cs" />
     <Compile Include="Game\RainMeadow.EntityHooks.cs" />
-    <Compile Include="RainMeadow.GameplayHooks.cs" />
+    <Compile Include="Game\RainMeadow.GameplayHooks.cs" />
     <Compile Include="RainMeadow.Logging.cs" />
     <Compile Include="Menu\RainMeadow.MenuHooks.cs" />
     <Compile Include="Online\Events\RequestResult.cs" />
